Boost: Use Headers instead of building when not required
When building USD without Python, OIIO or VDB, Boost libs aren't required.
In this scenario, it is preferable to just copy over the headers instead and save some time and complexity.

As requested, I have added an ignore list to bring down the size of the headers considerably.
Prior to this PR, boost would include ~170M-180M of headers. After this patch, boost only includes about 37M of headers. This could be further reduced in a future PR with better heuristics, but a 80% savings in size felt like a good win.

This also has a significant benefit that it allows the core USD build to be supported on platforms that Boost doesn't natively build on yet like iOS and visionOS. It means that the build_usd.py script doesn't need to carry as many patches to Boost itself if someone is just building the core USD library.

This is similar to [PR 2914](https://github.com/PixarAnimationStudios/OpenUSD/pull/2914) but can short circuit the entire bootstrap and b2 process, which may still error on platforms that b2 doesn't understand.