[usdImaging] Fix GeomModelAPIAdapter
An error in UsdImagingGomModelAPIAdapter caused the applyDrawMode property to be mislocated. We did not notice this because the code is only exercised by the draw mode standin when stage scene index is enabled. Currently, no draw mode tests are run with stage scene index enabled because the draw mode standin relies on geom subsets support, which is not yet implemented in hydra 2.0.

This change is a requirement for correct geom subset support.

(Internal change: 2316117)