Refactoring the implementation of Pcp_ComputeRelocationsForLayerStack to be more robust in preparation for adding validation code that checks for conflicting relocates within the layer stack.
This refactor also fixes a relocates bug where the origin source for
nested ancestral relocates would not always be computed correctly and
the success and failure would depend on the sort order of the paths
that were relocated. This could manifest as erroneous cycle detection
and incorrect mappings of target or connection paths across a relocation
node. Luckily TrickyMultipleRelocations4 already demonstrated this bug
and it has been updated to test the relationship target mapping issue.
It also no longer produces the erroneous cycle error mentioned.

(Internal change: 2316621)