[hdsi] Fixes handling of dirtied entries
A scene index may send out dirty prim entries with the "empty locator".  This indicates that everything on the prim is dirty.

The previous code was seeing if the locator contained the TetMesh schema to try ultimately update them to dirty the mesh schema.  If the locator has the empty locator, it contains everything so does not need any updating.

The previous code would then replace the locator with a new one, but this would discard any other locators.  This change fixes it to keep the locators that aren't the "double-sided" or "topology" schemas.

(Internal change: 2317028)