pcp: Reintroduce specializes culling optimization
This optimization was originally introduced in change 2312006
to fix memory usage increases from prior changes. It was
reverted in 2314652 due to runtime regressions in performance
tests for authoring workflows.

We've decided to restore this optimization after further testing
showed significant memory usage benefits on production shots
that heavily use specializes arcs. For example, memory usage
for one shot had previously increased from 1.7GB to 2.2GB. With
this optimization, memory usage dropped to 1.4GB -- down 18%
from the original value and 36% from the increased value.

The runtime performance issues showed up in a synthetic
scalability test that authored overrides via the USD API
on 10-20k attributes on prims using specializes arcs.
Each edit incurred an extra .12ms (or 120us) to recompose
a single prim index, which is not noticeable for interactive
workflows. While the extra time *is* noticeable in aggregate,
there are alternatives (such as dropping to Sdf to perform
authoring in a change block) to reduce this cost. Because of
this, we believe the memory benefits of this optimization
are far more beneficial for users.

(Internal change: 2317035)