Revise test case comparison parameters to avoid false positive
Problems:
As comparing test case output with baseline, the OIIO tool idiff consume a few options, e.g. -fail, -failpercent, -p (perceptual) and so on. From very begining, the options value inputs are not reasonble to catch all of pixel changes. The false positive is found when using WebGPU backend. That is, a number of cases were passed, but rendering images are not correct. The results are not real positive.

Solutions:
Change test input parameters to be stricter in CMake file to successfully detect errors in
the generated images. Another side, remove some perceptual options they are not proper to test for specific cases.