pcp: Fix culling bug with hierarchies of specializes
Nodes that could have been culled were being missed in cases
involving hierarchies of specializes arcs (e.g. A specializes
B specializes C). This was due to the strong-to-weak iteration
when culling propagated specializes nodes.

In the example above, the prim index would contain a chain of
nodes A -> B -> C, where B and C are specializes nodes. These
would be propagated to the root of the prim index as siblings;
i.e. the root node would have two child specializes nodes for
B and C in that order. The strong-to-weak ordering meant we
would try to cull B in the original chain of nodes first,
but that would fail because C had not yet been culled, and a
node can only be culled if all its children are also culled.

The fix reverses the iteration order to avoid this problem.

(Internal change: 2317401)