Always add sdrUsdDefinitionType metadata for all sdrShaderProperties parsed from UsdShadeShaderDef
Shader Definitions in UsdShadeShader already have their properties
conforming to SdfValueTypeNames so:
in order to retain this information, we save these types in
sdrUsdDefinitionType we do not parse any explicitly specified
"sdrUsdDefinitionType" in a UsdShadeShaderDef, since the property types are
already conforming to SdfValueTypeNames.

Other updates in this change are in accordance to the above main change, do
note some important updates here though:
- OSL shader for UsdUVTexture is updated to include sdrUsdDefinitionType for
wrapS and wrapT which should have been token to begin with, so as to conform
to the UsdShadeShaderDef specification.
- Do note the removal of boolButInt properties in some tests, as these are
already conforming so are not needed. + we do not have any explicit role for
sdrUsdDefinitionType in UsdShadeShaderDef
- Note that special logic for token/string in UsdLuxLight test is updated in
as this is no longer an exception!

(Internal change: 2318704)