tf: use MSVC preprocessor version to determine which set of macros to use in preprocessorUtilsLite.h
In the cases where we were previously checking whether the compiler was MSVC
when deciding whether to use an alternative set of macros versus other
toolchains, what we really wanted to know was whether MSVC's "traditional"
preprocessor was being used.

This replaces the use of ARCH_COMPILER_MSVC in preprocessorUtilsLite.h with the
newly added ARCH_PREPROCESSOR_MSVC_TRADITIONAL so that only in cases where we're
compiling with MSVC *and* using its traditional preprocessor do we use the
alternate macros. Otherwise when using MSVC's newer, more standards-conforming
preprocessor, we use the same macros as other toolchains.