[usdImaging] Legacy adapter hijacking improvements
Due to adapter hijacking, HdInstanceAdapter, HdPointInstancerAdapter, and HdDrawModeAdapter receive calls from the scene delegate for information about the underlying prims they mask. They maintain internal data structures to be able to forward these requests to the appropriate prim adapters. In all three cases, the adapters were updating these internal structures *after* inserting the masked prims into the render index. This means that they would be unable to respond to requests that came in as a direct consequence of insertion.

This is exactly what happens during emulation. When the adapters insert the masked prims into the render index proxy, the proxy immediately sends the prim on to HdLegacyPrimSceneIndex for adding. HdLegacyPrimSceneIndex therefore was unable to use any information other than path and type to make decisions about *how* to insert the prim. The scene delegate could retrieve no information about prims masked by the hijacking adapters at that time.

This change reorders the population of internal data structures in the hijacking adapters so that these structures are fully populated *before* the masked prims are added to the render index. This is especially tricky in the case of the instancing adapters, which relied on the insertion call's return of the inserted prototype prim path. Therefore, a mechanism for precomupting the prototype prim's insertion path is required. This has been achieved by hoisting ResolveCachePath, which was previously protected and only on UsdImagingGPrimAdapter and UsdImagingLightAdapter, to a public method on UsdImagingPrimAdapter, and by implementing a custom override for UsdImagingPointInstancerAdapter. Since UsdImagingGPrimAdapter's implementation was used by many instanceable non-gprims, it has been moved to an intermediate abstract adapter class, UsdImagingInstanceablePrimAdapter. UsdImagingGPrimAdapter and other instanceable non-gprim adapters now derive from this class instead of UsdImagingPrimAdapter.

Note that this required removing the const qualifier from the following methods on the UsdImagingPrimAdapter interface, since some adapters' implementations of these methods need to be able to call ResolveCachePath, which is non-const:
* GetExtComputationInputs
* GetExtComputationPrimvars
* GetExtComputationInput
* TrackVariability

(Internal change: 2318924)