HdxTaskController: fixing code path to upadte clearValue for aov bindings.
There was an optimization to check whether the existing clear value was equal to the desired clear value before setting the clear value.

However, the clear value that was set was different from the clear value that was tested against. If the set of tasks changed, this could potentially lead to not setting/or unsetting the empty value as clear value on a task.

(Internal change: 2319089)