[Hd] HdSceneIndexPluginRegistry: Adjust the order of how scene index plugins concatenate, making phase the most significant ordinal component.
Previously, all renderer-agnostic phases were appended,
followed by all renderer-specific phases.  As a result,
there was no way for a renderer-specific plugin to
run before a renderer-agnostic one, regardless of their
relative phase.

To enable filtering generative procedurals in specific
renderers, prior to the index that expands them for
all renderers, we adjust this logic so that phase is
the major component of the ordering.  Within a phase,
renderer-agnostic scene indexes still continue to
precede renderer-specific ones.

(Internal change: 2320013)