Added attribute type checking on shaders to UsdChecker
This check retrieves the Sdr type of all the inputs of the shader node and converts it to the respective Sdf type. These types are then used as the ground truth when validating the attribute type.

Added test case badPropertyType.usda that contains properties with incorrect Sdf types.

Also updated "varname" from token to string type to make it type compliant with the specification so that testUsdUtilsCreateNewUsdzPackage and testUsdUtilsCreateNewARkitUsdzPackage pass with this new type checker.

Fixes #2729

(Internal change: 2320085)