[arch] Arch_DebuggerIsAttachedPosix: use TracerPid in /proc/self/status
The old implemenation worked by trying to attach a ptrace to the current
process, and if it failed, then we assumed it was being debugged.

The problem is that we have no easy way to check if an EPERM result
from ptrace is because it's being debugged, or an actual permissions
error - and permissions checking process is somewhat complex - see the
"Ptrace access mode checking" section here:

https://man7.org/linux/man-pages/man2/ptrace.2.html

On a basic Ubuntu-22.04 test system, ptrace would return EPERM even when
the process was not being debugged, resulting in a "false positive".
The new TracerPid method was more accurate in this case.

Checking the /proc/self/status for TracerPid is also the same method
that gperftools and even gdb itself use to check if a process is being
debugged:

- https://github.com/gperftools/gperftools/blob/02adc8ceab39bbeac1f65e10bde577e1753094fa/src/heap-checker.cc#L183
- https://sourceware.org/git/?p=binutils-gdb.git;a=blob;f=gdb/nat/linux-procfs.c;h=b17e3120792e0e0790271898212b69b0577847cc;hb=HEAD#l68
- https://sourceware.org/git/?p=binutils-gdb.git;a=blob;f=gdb/testsuite/gdb.threads/siginfo-threads.c;h=22c6038206ba77fc3432da5ee30284c80641f305;hb=HEAD#l373