Fixing issue where subroot references to namespace children under of a payloaded prim were failing to automatically load the ancestral payload when the reference is internal.
The issue was that the way we determined if a payload needed to always
be loaded was to attempt to map its path all the way up to the root of
the prim index we're constructing. If the path fails to map all the
way, we assume that the payload belongs to an ancestral prim index that
we're building for a subroot arc and force load the payload.

However this breaks down if the subroot arc that introduces the
ancestral payload is an internal reference arc. Internal references
always have an additional identity mapping (that is not present in
non-internal references) that allows the ancestral payload's path to map
to the root as itself and the payload will only be loaded if the path is
requested.

The fix is change how we map payload inclusion paths to the root to
specifically ignore identity mappings that come from internal reference
(and payload) arcs so that we have the same subroot ancestral payload
inclusion behavior for both internal and non-internal subroot references.

(Internal change: 2321525)