Get motion blur to work with usdrecord
In order for motion blur to work with usdrecord, you need a
valid renderSettings prim with a relation to the camera with
the shuter interval set, and and to set both the camera
and frame number in the usdrecord arguments.
Note that if the camera is the 'main_cam' usdrecord will
default to using that camera so we do not need to set the
relationship on the renderSettings prim or add the camera
argument.

There were two main issues preventing this to work.
The first is that the only way to get the correct shutter interval
to Renderman is via a valid RenderSettings prim. Meaning we
need to author a renderSettings prim with a render product
and render var. We do not however need to set the
'renderSettingsPrimPath' in the stage metadata or specify
the path with the usdrecord argument.

The second issue is due to the default start time value used for
usdrecord. When no frames are listed and the --defaultTime
argument is not given, then the start time used is EARLIEST
which causes any sample interval to be trivial so the prims
are sampled just once and we loose get any motion blur.

(Internal change: 2321528)