Making the behavior for legacy relocates more consistent when encountering the condition of multiple relocates from different sources to the same target path. This condition is erroneous but the legacy behavior still processed these relocates (the new behavior rejects all relocates with the same target when this condition is encountered).
In this case, the use of an unordered map made the "winner" of the
target to source mapping prone to change based the state of the SdfPath
pool. The fix is to make sure we're choosing the "winning" target to
source mapping based on the source's SdfPath lexicographical order so
that the behavior is always the same regardless of the SdfPath pool
state.

(Internal change: 2321938)