PcpMuseum tests now can be (and now are) run in USD mode to test both USD and non-USD behaviors.
This involved adding a --usd flag to testPcpCompositionResults that
tells the test to create its cache in USD mode. All the PcpMuseum tests
now run testPcpCompositionResults in both modes and compare the results
against the same baselines in all but a handful of cases where the USD
mode behavior is different (unique USD baselines are used for these
exceptions). Tests involving relocates have their USD mode comparisons
temporarily disabled until relocates in USD mode are actually enabled.

This also required some changes to diagnostics and API that relate to
property targets and prim stacks as USD mode doesn't cache some of this
information like non-USD mode does. In USD mode we compute these things
on the fly when needed so we can still produce the same output in the
museum tests.

(Internal change: 2322165)