Fixing memory regressions due to unnecessarily creating and storing relocation map expressions when we know there are no relocates on a layer stack.
This also required an update to Pcp change processing as the relocation
map expressions were used to update map functions in prim index nodes
automatically when relocates change without having to rebuild all prim
indexes. But since these map expressions are not present when we have no
relocates, we have to make sure we resync everything that depends on a
layer stack when an edit adds relocates to a layer stack that originally
had none.

(Internal change: 2322674)