[usdImaging] Promote instancer prototypes from properties to child prims
In legacy UsdImaging, imageable prims in an instancer prototype were sent to Hydra as properties of their parent instancer. This prevented the imageable prototype prims from having child subprims. As a result, any geom subsets in the prototype prims were also attached as properties. Since geom subsets are in many ways similar to imageable prims, Hydra requires specialized handling for geom subsets with respect to such concepts as visibility and material bindings. This specialized handling was in many cases duplicative and bothersome to maintain. Furthermore, in Hydra 2.0, geom subsets transit the scene index as prims, not as properties. This means the Hydra 2.0 scene indices must keep all that special handling for legacy property-type geom subsets to maintain compatibility when front-end emulation is in use. If instancer prototype prims could take geom subsets as child prims, their path through Hydra could be greatly simplified. But they cannot so long as the prototype prims are inserted as properties of the instancer.

This change promotes the prototype prims from properties to child prims. This will allow the prototype prims to themselves take geom subsets as child prims. A forthcoming change will unify the path through Hydra for geom subsets and remove the redundant handling.

(Internal change: 2322974)