[usdviewq] Add shortcut Ctrl+Shift+R to reload stage.
Reloading the stage is a useful workflow action when hand-authoring USD test assets to see changes in the rendered image.
Due to certain limitations in UsdImaging 1.0's change processing, it doesn't suffice to reload all layers (currently bound to Ctrl+R). Reloading the stage results in the imaging system getting reinitialized and should provide the same result as bouncing the app.

(Internal change: 2323992)