[usdImaging] Fix for missing cards textures under instancing
In instancing topologies where a point instancer prototype is a native instance, draw mode failed, typically presenting as missing cards textures. The point instancer adapter was erroneously sending the the rerooted native prototype to the draw mode adapter for population instead of the original prototype prim. This has been corrected. This case was not covered by any existing regression test, so the existing test testUsdImagingGLCards has been extended to cover it.

(Internal change: 2324397)