pypi: Pin delocate package to 0.10.x
The delocate package used to construct the usd-core wheel on
MacOS was recently updated to 0.11.0. This version introduced
new behavior that modified the name of the generated wheel,
which caused issues later in our process. This behavior has
affected other projects; see:

https://github.com/matthew-brett/delocate/issues/214.

In the meantime, we pin delocate to the prior 0.10.x version
to get things working again. Support for Python 3.6 was dropped
in 0.10.3, so we use a "compatible release" version specifier
that gets us 0.10.2 for Python 3.6 and the latest 0.10.x version
(currently 0.10.7) for Python 3.7+.

(Internal change: 2324479)