UsdImagingDataSourceCustomPrimvars: do not store attribute queries in map but create it every time a primvar is querried.
Without this, a primvar would not become visible the first time a value is authored if a client were to store the primvars container.

Also, if the attribute query might contain stale resolution info resulting in the wrong value. So we need to make sure that we recreate the attribute query.

Note that we store the attribute query in the primvar data source, not the primvar value data source. This is ok because we send out dirty locators for the primvar rather than the primvar value in response to the USD invalidation.

(Internal change: 2324493)