[usdImaging] Fix up PointInstancer primvar semantics.
This change adds "angularVelocities" support to point instancers in UsdImagingDelegate; "velocities, accelerations, angularVelocities" support in UsdImagingStageSceneIndex, and support for local constant/uniform primvars and inherited constant primvars to show up in hydra as "constant" primvars on a point instancer.

This should bring the usdImaging point instancer support in line with the UsdGeomPointInstancer spec, and address issues in Solaris that were proposed to be fixed in PRs #1731 and #1734.  This change includes and expands on those PRs.

Fixes #1731
Fixes #1734

(Internal change: 2324658)