Usdview Prman sceneLight behavior
This change adds the DirtyResource bit when switching en/disabling scene lights
It also makes sure that we are always calculating the HdMaterialNetwork
for the light, and that when scene lights are disabled we overrite the
attribute value for intensity to 0 before returning it to the light.

This change also fixes a bug in the primTypePruningSceneIndex to make sure
that when pruned, prims are properly removed.

(Internal change: 2324819)