arch/stackTrace: fix a formatting bug and a compiler error.
When a negative number was passed to asitoa(), the minus sign was always absent,
overwritten by the first digit.  The issue and the fix are not easy to verify
with existing test cases, but I verified both via separate compilation and
testing.

Confusingly, this also caused gcc to sometimes report that we could be writing
off the beginning of the output buffer, which in strict mode would break the
build.  I'm fairly sure this was never the case, but gcc seems placated with
this fix.

(Internal change: 2325441)