Hydra render pass work
Add a RenderPassSceneIndex which applies the active render pass
to the scene.  For now, this lives in HdPrman because it includes
Prman-specific concepts like matte-ing.  In the future we may
factorsome of this to live in Hdsi for use in other renderers.

This is an initial checkin that provides evaluation of render
passes and a unit test, but not invalidation..  A subsequent change
will add invalidation and change processing for interactive
workflow support.

(Internal change: 2325598)
(Internal change: 2325637)