HdPrman now supports field-less RiVolume primitives, such as those used for homogeneous volumes (which do not require any voxel data).
These volumes require an Ri:type parameter (Prman supports
box, ellipsoid, cone, cylinder) and typically also Ri:Bounds.
To support Ri:Bounds, update the primvar conversion code to allow
the 6-component Ri:Bounds as a constant-variability Hydra primvar.

Volumes are typically used with RenderMan for fog because
it supports putting the camera inside an RiVolume but does
not support putting the camera inside a mesh (ex: sphere)
with volume shader.  The underlying intuitive reason is that
it does not have a way to determine point-containment inside
a mesh, so it cannot properly initialize integration of paths
in the case of camera inside a volumetric mesh; whereas an
RiVolume has trivial bounds, making this an easy query.

(Internal change: 2326572)