[hdPrman] Use group:membership for renderTag-based visibility
hdPrman's prior approach to updating visibility in respose to changes in included purpose would make calls to ModifyGeometryInstance that only included the visibility parameters, causing all other parameters to fall off.

The new approach uses a feature of PbsPathTracer to disable tracing of geometry assigned to certain trace groups. We assign each geometry instance to a group based on its computed purpose. During renderpass execution, we traverse the rprims in the scene to gather up all known purposes, subtract the active purposes, and tell PbsPathTracer to hide the remainder.

Support for this approach will be extended to PxrPathTracer in the near future.

(Internal change: 2327209)