[hd] Improve memory usage in HdLegacyGeomSubsetSceneIndex
Previously, HdLegacyGeomSubsetSceneIndex maintained a list of all subsets for every mesh or basisCurves prim added to the scene, regardless of whether the prim had any subsets. With this change, only those prims that actually have subsets are tracked. This saves considerable memory on scenes with very many mesh or basisCurves prims.

(Internal change: 2327343)