[hgiVulkan] Rather than flipping the Vulkan viewport in the y-direction, instead allow the geometry to render upside down, which happens naturally when using an OpenGL projection matrix with Vulkan. Typically rendering in Vulkan with an OpenGL projection matrix would result in upside down geometry that is also facing the wrong way, so to combat the undesired latter effect, we flip the winding order in HgiVulkan.
Rendering upside down in HgiVulkan allows us to avoid a number of changes that
would otherwise be needed to get the correct result, due to the current
OpenGL-styling of Storm. For example, discrepancies between framebuffer
coordinates (y-down in Vulkan versus y-up in GL), such as when doing an AOV
readback in the shader, now don't need any special handling.

(Internal change: 2327844)