[usdImaging] Add GetImagingSubprims to RepresentedByAncestorPrimAdapter.
Note that the fallback warning in PrimAdapter::GetImagingSubprims is meant to identify adapters that have not been updated for UsdImaging 2.0.  That warning isn't appropriate for RepresentedByAncestorPrimAdapter, which does occasionally need to answer the question of what subprims it has due to nonsensical queries like GetChildPrimPath("/path/to/shader").  The answer, of course, is that it has no subprims.

(Internal change: 2328867)