Build with MaterialX v1.38.10
Control over the dome light intensity was added to v1.38.10
so dome light intensity was added to some of the test cases in
testUsdImagingGLMaterialXBasic to impove the basline visibility.
Doing this however highlights the existing visual differences
between between local and gauntlet runs particularly when it
comes to the specular highlights.

(Internal change: 2329217)