[hdx, hdSt] Certain Hdx task types now internally retain the GLSLFX form of their shaders, to avoid re-parsing them each time a task is constructed.
Profiling RenderMan render startup time in a production shot
revealed non-trivial time spent loading GLSLFX files as part
of setting up the surrounding Hydra Hdx presentation pipeline.
Using TF_DEBUG set to HIO_DEBUG_GLSLFX showed 3,506 calls
to HioGlsfx to load a shader from the fileystem, but only
18 unique shaders.  These 18 shaders are shipped with
the code, not production assets.

Storm (aka HdSt) does have a resource registry that provides
re-use of GLSLFX shaders. However, that is currently unavailable
to Hdx, because the latter is not part of Storm, but instead
part of the surrounding render task system, and it does not
have access to the Storm resource registry.  In particular, when
Renderman (aka HdPrman) is the chosen render delegate, it provides
its own resource registry, which does not provide GL shadertext
caching.

In the long run this seems worth revisiting to provide
a more robust way to avoid re-parsing GLSLFX files;
in the short term we introduce a retained internal
HioGlslfx form of these shaders so that Hdx tasks may
share and reuse them.

Support the Hdx change, HdStRenderPassShader now accepts
an HioGlslfx object representing a pre-parsed shader,
as well as the glslfx filename as before.

(Internal change: 2329387)