[usdviewq] Update OCIO colorspace selection logic.
Previously, the handler for the OCIO "display/view" menu would attempt to set the colorspace as well, using getDisplayViewColorSpaceName.  However, this is incorrect; it's better to leave the colorspace as-is.  Note that the default value ("") that gets passed in if no colorspace is explicitly set will be expanded by HdxColorCorrectionTask into a reasonable default (taking into account display/view selection).

Thanks to Doug Walker and Ken McGaugh for the tip!

Fixes #3017

(Internal change: 2330008)