[hd] Fix an assert when calling prim.dataSource->GetNames() on a legacy geom subset prim.
This only came up in the context of the hydra scene browser, but calling DataSourceLegacyPrim::GetNames() for subset prims isn't safe because that will call _IsLight, which calls delegate->GetLightParamValue, which will fail to find the subset prim path in _primInfo, since the subset prim was added by the scene index.  Instead, GetNames() is just aligned to the local implementation of Get().

(Internal change: 2330013)