[hdPrman] Add env setting HDPRMAN_TESSELLATE_IMPLICIT_SURFACES, defaulting to false.
hdPrman supports a few implicit shapes, which we try to pass from USD as unaltered as possible, but these have different imaging behavior from the legacy tessellations we've been using.  Even though the hdPrman interpretations are more correct, it's useful to be able to force tessellation to compare UsdImaging implementations.

(Internal change: 2330021)