usdShade: implement and register the "SubsetMaterialBindFamilyName" validator
This validator checks whether a geom subset has authored material bindings
but no authored subset family name, in which case an error is emitted
indicating that the family name should be set to "materialBind" to ensure
that the material bindings are visible to renderers. The material bindings
will have no effect otherwise.