usdShade: implement and register the "SubsetsMaterialBindFamily" validator
This validator checks that if an Imageable prim has GeomSubsets of the
"materialBind" family, the family must be of a restricted type (either
"nonOverlapping" or "partition"), since it is invalid for an element of
geometry to be bound to multiple materials. Those subsets are also checked
to ensure that they are of element type "face", since material bindings may
only be applied to geometric faces.