[hd] Make HdRenderDelegate::IsParallelSyncEnabled virtual.
The intent with this function is to let render delegate authors gradually opt into multithreaded sync for sprim and bprim types (such as materials).  HdRenderIndex::SyncAll will dispatch sprims to a threadpool if this function returns true for a given type, but for most types it currently still returns false, as many render delegates have thread hazards in sprim and bprim sync.

(Internal change: 2330778)