cmake: Fix bug that caused oneTBB to not be found on Windows
The build system finds the TBB package via this call in Packages.cmake:

  find_package(TBB REQUIRED COMPONENTS tbb)

On Windows oneTBB names its shared library "tbb12.dll" and "tbb12.lib".
FindTBB.cmake was updated to try to accommodate this, but the way it
was done required the component passed to find_package to be changed
to "tbb12" instead of "tbb" on Windows.

This change reverts that update and instead translates the "tbb"
component to "tbb12" within FindTBB.cmake when it calls find_library
to find the individual shared libraries for the package. This allows
us to keep the find_package call unchanged.

(Internal change: 2331094)