[hdSt] Set default compute shader local size to (64, 1, 1) in HdSt_CodeGen::_CompileWithGeneratedHgiResources. This codepath is used by ext GPU computations and the frustum culling compute pipelines. Previously they were using a local size of (1, 1, 1).
To ensure correctness, we add a variable "indexEnd" that is the size of
the total number of computations/compute shader invocations to run to the end
of the shader's constant params. The shader will early out if the
computeCoordinate is greater than or equal to this value (this is similar to
what we do for other compute pipelines in Storm, such as quadrangulation or
subdivision).

(Internal change: 2331615)