HdxSelectionSceneIndexObserver::SetSceneIndex() now populates the initial selection in parallel.  As used by HdxColorizeSelectionTask::Prepare in a production shot, this took 1500 msec previously, now ~260 msec.
The bulk of the work is traversing the input scene invoking
GetPrim() in order to query selection schemas.  Although the host
app may have a centralized selection object, the Hydra2 scene index
design distributes this state to selection schemas on individual
prims, requiring a gather step if a single collection is needed
downstream, as in HdxColorizeSelectionTask.  One benefit of the
per-prim schema approach is that selection state remains coherent
with prims even if they are re-rooted or pruned by a scene index.

(Internal change: 2331793)