usdviewq, testusdview: Refactor to reduce duplication
Usdviewq.Launcher.__LaunchProcess was previously not
intended to be overriden and used Python's name mangling
to try to prevent this. This forced the TestUsdView
launcher to override the Run method. The implementation
duplicated option handling to try to match Usdviewq's
behavior, but failed to do so in all cases.  In
particular, the "--traceToFile" option was not handled
properly, so generating traces when using testusdview
did not work.

This change turns __LaunchProcess into an overridable
LaunchProcess function, along with some other small
refactorings. This allows TestUsdView to just override
certain behaviors instead of the whole Run function,
which removes code duplication and fixes the issue
above.

(Internal change: 2331981)