Adding "relocate to delete" feature to Pcp by handling relocates whose target is an empty path.
Technically "relocate to delete" has already existed since relocating a
prim to a path where the parent doesn't actually exist would effectively
delete the relocated prim. This behavior is now formalized by allowing
relocates to the empty path to make it explicit that the intention of
the relocate is to delete the prim from its parent.

(Internal change: 2331999)