[hdPrman] Code fixes/enhancements for lights
Add support for PxrEnvDayLight and some bug fixes with mesh and volume lights.

We also add a preprocessor define, HD_PRMAN_HIDE_DEFAULT_DOMELIGHT_TEXTURE, which forces textures of any "default" domelights (with a </_UsdImaging_> or </husk_headlight> prefix) to be invisible. This is a temporary measure to support customers transitioning to using scene indices.

(Internal change: 2332992)